requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
streamlit>=1.40.0
plotly>=5.24.0
pandas>=2.0.0
tweepy>=4.14.0
requests>=2.31.0
